--- src/Zebra.hs
@@ -1,7 +1,0 @@
-{-# LANGUAGE NoImplicitPrelude #-}
-module Zebra (
-    module X
-  ) where
-
-import           Zebra.Data as X
-import           Zebra.Serial as X
