+++ src/Zebra/Data.hs
@@ -6,3 +6,4 @@ module Zebra.Data (
 import           Zebra.Data.Block as X
 import           Zebra.Data.Core as X
 import           Zebra.Data.Encoding as X
+import           Zebra.Data.Fact as X
