+++ src/Zebra/Data/Entity.hs
@@ -17,7 +17,7 @@ import           P
 import           X.Text.Show (gshowsPrec)
 
 import           Zebra.Data.Core
-import           Zebra.Data.Table
+import           Zebra.Table
 
 
 data Attribute a =
