+++ src/Zebra/Foreign/Table.hs
@@ -25,9 +25,9 @@ import           P
 
 import           X.Control.Monad.Trans.Either (EitherT, left)
 
-import           Zebra.Data.Table (Table(..), Column(..))
 import           Zebra.Foreign.Bindings
 import           Zebra.Foreign.Util
+import           Zebra.Table (Table(..), Column(..))
 
 
 newtype CTable =
