+++ test/Test/Zebra/Table/Mutable.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE TemplateHaskell #-}
-module Test.Zebra.Data.Table.Mutable where
+module Test.Zebra.Table.Mutable where
 
 import           Control.Monad.ST (runST)
 
@@ -22,11 +22,11 @@ import           Text.Show.Pretty (ppShow)
 
 import           X.Control.Monad.Trans.Either (runEitherT)
 
-import           Zebra.Data.Fact (Value)
-import           Zebra.Data.Schema (Schema)
-import           Zebra.Data.Table (Table, ValueError)
-import qualified Zebra.Data.Table as Table
-import qualified Zebra.Data.Table.Mutable as MTable
+import           Zebra.Schema (Schema)
+import           Zebra.Table (Table, ValueError)
+import qualified Zebra.Table as Table
+import qualified Zebra.Table.Mutable as MTable
+import           Zebra.Value (Value)
 
 
 prop_default_table_vs_mtable :: Property
